refractor: 홈뷰 리팩토링
