+++ Plantory/Plantory/ViewModels/Chat/CalendarEntry.swift
@@ -7,9 +7,6 @@
 
 import Foundation
 
-//더미데이터
-struct CalendarEntry {
-}
 
 
 //홈화면 캘린더용 일기 더미데이터
